+++ src/main/java/com/floriculturamonteiro/floricultura/service/AdminService.java
@@ -1,14 +1,8 @@
 package com.floriculturamonteiro.floricultura.service;
 
-import com.floriculturamonteiro.floricultura.model.Carrinho;
-import com.floriculturamonteiro.floricultura.model.Cliente;
 import com.floriculturamonteiro.floricultura.model.Flores;
-import com.floriculturamonteiro.floricultura.repositories.CarrinhoRepository;
-import com.floriculturamonteiro.floricultura.repositories.ClienteRepository;
 import com.floriculturamonteiro.floricultura.repositories.FloresRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Example;
-import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
@@ -19,9 +13,6 @@ public class AdminService {
 
     //atributo
     private final FloresRepository repository;
-    private final CarrinhoRepository carrinhoRepository;
-    private final ClienteRepository clienteRepository;
-
 
     //adicionar as flores
     public void addFlor(Flores flores){
@@ -81,32 +72,5 @@ public class AdminService {
     public List<Flores> getAllFlores(){
         return repository.findAll();
     }*/
-
-    //exibir as compras no controle de vendas
-    public List<Carrinho> exibirCompras() {
-        return carrinhoRepository.EncontrarTodosComItens();
-    }
-
-    public List<Carrinho> pesquisarCliente(String nome){
-        if (nome == null || nome.isBlank()) {
-            return exibirCompras();
-        }
-        return carrinhoRepository.findByNomeClienteIgnoreCase(nome);
-    } //buscar carrinho pelo nome do cliente no controle de vendas
-
-    public List<Carrinho> pesquisaByExample(String nome){
-        Cliente cliente = new Cliente();
-        cliente.setNome(nome);
-
-        ExampleMatcher matcher = ExampleMatcher
-                .matching()
-                .withIgnoreCase()
-                .withStringMatcher(ExampleMatcher.StringMatcher.CONTAINING);
-
-        List<Cliente> clientes = clienteRepository.findAll(Example.of(cliente, matcher));
-
-        return carrinhoRepository.findByClienteIn(clientes);
-
-    } //abordagem de pesquisa alternativa utilizando a interface QueryByExample
 
 }
